weird crashing but reworked a lot

+++ data/include/mathhandler.h
@@ -1,4 +1,6 @@
-int calcSNMPPulses(unsigned long Avg);
-CRGB calcPulseColor(unsigned long Avg);
-CRGB calcPulseColor2(unsigned long Avg);
+#include <FastLED.h>
+
+int calcSNMPPulses(int Avg);
+CRGB calcPulseColor(int Avg);
+CRGB calcPulseColor2(int Avg);
 int randomZeroOrOne();
